duplicated ./lib/
...so that we are not forced to keep the library
for the two apps in sync.

Actually there is no need yet to let the libs diverge.
But there were TS problems due to common dependencies of the lib and
the apps.  I could have solved this by moving node_modules/ to the
top-level, but that would have coupled the apps even more tightly.